move another extension class into own file
